add support for 16bit positions values

+++ src/replicatorg/drivers/gen3/Makerbot4GDriver.java
@@ -99,11 +99,11 @@ public class Makerbot4GDriver extends Sanguino3GDriver {
 		PacketBuilder pb = new PacketBuilder(MotherboardCommandCode.SET_POSITION_EXT.getCode());
 
 		Point5d steps = machine.mmToSteps(p);
-		pb.add32((long) steps.x());
-		pb.add32((long) steps.y());
-		pb.add32((long) steps.z());
-		pb.add32((long) steps.a());
-		pb.add32((long) steps.b());
+		pb.add32((int) steps.x());
+		pb.add32((int) steps.y());
+		pb.add32((int) steps.z());
+		pb.add32((int) steps.a());
+		pb.add32((int) steps.b());
 
 		Base.logger.log(Level.FINE,"Set current position to " + p + " (" + steps
 					+ ")");
